+++ config/config.exs
@@ -14,8 +14,9 @@ config :collaboration,
   ga_dev_code: "UA-119119225-1",  # Google Analytics Dev Code
   ga_prod_code: "UA-119138942-1", # Google Analytics Production Code
   minTime: 10 * 60,               # minimal experiment time [sec]
-  passcode: System.get_env("PASSCODE"),           # passcode for experiment
-  password: System.get_env("PASSWORD_DEFAULT"),   # admin user password
+  # passcode: System.get_env("PASSCODE"),           # passcode for experiment
+  password: System.get_env("PASSWORD_DEFAULT"),   # admin user password,
+  survey_link: "https://usf.az1.qualtrics.com/jfe/form/SV_cUdM8cbGDj6tnvf",
   delayed_likes: %{
     1 => [],
     2 => [],
@@ -25,16 +26,6 @@ config :collaboration,
     6 => [{  1, 530 }, {  2, 180 }, {  6, 90 }],
     7 => [{ 12, 530 }, { 13, 180 }, { 18, 90 }],
     8 => [{  1, 530 }, {  2, 180 }, {  6, 90 }]
-  },
-  survey_codes: %{                # needed for survey upon finishing experiment
-    1 => "condition 1 code",
-    2 => "condition 2 code",
-    3 => "condition 3 code",
-    4 => "condition 4 code",
-    5 => "condition 5 code",
-    6 => "condition 6 code",
-    7 => "condition 7 code",
-    8 => "condition 8 code"
   }
 
 # Configures the endpoint