+++ lib/collaboration_web/controllers/user_controller.ex
@@ -1,13 +1,15 @@
 defmodule CollaborationWeb.UserController do
   use CollaborationWeb, :controller
 
+  alias Collaboration.Accounts.User
+
   # start page to create experiment users
   def new(conn, _params) do
-    render conn, "new.html", changeset: change_user()
+    render conn, "new.html", changeset: User.changeset(%User{})
   end
 
   def create(conn, %{"user" => user_params }) do
-    case create_participant(user_params) do
+    case create_user(user_params) do
       {:ok, user} ->
         conn
         |> CollaborationWeb.Auth.login(user)