+++ lib/collaboration_web/views/session_view.ex
@@ -1,4 +1,6 @@
 defmodule CollaborationWeb.SessionView do
   use CollaborationWeb, :view
 
+  def survey_link(), do: Application.fetch_env!(:collaboration, :survey_link)
+
 end